+++ CHANGELOG.md
@@ -7,7 +7,12 @@
   - `live_patch` - deprecate in favor of new `<link patch={..}>` component of `Phoenix.LiveView.Helpers`
   - `push_redirect` - deprecate in favor of new `push_navigate` function on `Phoenix.LiveView`
 
-## 0.17.10 (Unreleased)
+### Enhancements
+  - [Helpers] Add `dynamic_tag` function component
+  - [Helpers] Add `link` function component
+  - Add new attrs `:let` and `:for`. We still support `let` but the Formatter will convert it to `:let` and soon it will be deprecated.
+
+## 0.17.10 (2022-05-25)
 
 ### Bug fixes
   - [Formatter] Preserve single quote delimiter on attrs
@@ -20,7 +25,6 @@
 
 ### Enhancements
   - [Formatter] Introduce special attr phx-no-format to skip formatting
-  - Add new attrs `:let` and `:for`. We still support `let` but the Formatter will convert it to `:let` and soon it will be deprecated.
 
 ## 0.17.9 (2022-04-07)